tests moved to an interface
